+++ backend/jobs/cron.go
@@ -22,6 +22,7 @@ var (
 	checkMediuxSiteLinkJobID             cron.EntryID = 0
 	checkForMediaItemChangesJobID        cron.EntryID = 0
 	handleTempIgnoredItemsJobID          cron.EntryID = 0
+	refreshAnidbMappingsJobID            cron.EntryID = 0
 
 	// Configurable
 	autodownloadJobID cron.EntryID = 0
@@ -97,6 +98,8 @@ func GetListOfJobs() []JobInfo {
 				jobInfo.JobName = "Check for Media Item Changes Job"
 			case handleTempIgnoredItemsJobID:
 				jobInfo.JobName = "Handle Temp Ignored Items Job"
+			case refreshAnidbMappingsJobID:
+				jobInfo.JobName = "Refresh AniDB Mappings Job"
 			case kometaImportJobID:
 				jobInfo.JobName = "Kometa Asset Import Job"
 			default:
@@ -128,6 +131,8 @@ func TriggerJob(jobName string, jobID string) error {
 		entryID = checkForMediaItemChangesJobID
 	case "Handle Temp Ignored Items Job":
 		entryID = handleTempIgnoredItemsJobID
+	case "Refresh AniDB Mappings Job":
+		entryID = refreshAnidbMappingsJobID
 	case "Kometa Asset Import Job":
 		entryID = kometaImportJobID
 	default: